+++ packages/calendar/src/features/travel-calendar/components/TravelDateSingleTextInputField.tsx
@@ -14,6 +14,8 @@ export interface TravelDateSingleTextInputFieldProps {
   label?: string;
   onFocus?: () => void;
   calendarSize: TravelCalendarSizeVariant;
+  placeholderWhenBlurred: string | undefined;
+  valueWhenBlurred: string | undefined;
 }
 
 export const TravelDateSingleTextInputField: React.FC<
@@ -25,6 +27,8 @@ export const TravelDateSingleTextInputField: React.FC<
   localeCode,
   onFocus,
   calendarSize,
+  placeholderWhenBlurred,
+  valueWhenBlurred,
 }) => {
   const { mask, placeholder } = useMemo(() => {
     const dateFormatForLocaleCode = getDateFormatForLocaleCode(localeCode);
@@ -50,6 +54,8 @@ export const TravelDateSingleTextInputField: React.FC<
         label={label}
         placeholder={placeholder}
         calendarSize={calendarSize}
+        placeholderWhenBlurred={placeholderWhenBlurred}
+        valueWhenBlurred={valueWhenBlurred}
       />
     </Row>
   );