+++ packages/calendar/src/features/travel-calendar/components/TravelDateTextInputFields.tsx
@@ -18,6 +18,10 @@ export interface TravelDateTextInputFieldsProps {
   endDateLabel?: string;
   onFocus?: () => void;
   calendarSize: TravelCalendarSizeVariant;
+  placeholderWhenBlurredStartDate: string | undefined;
+  placeholderWhenBlurredEndDate: string | undefined;
+  valueWhenBlurredStartDate: string | undefined;
+  valueWhenBlurredEndDate: string | undefined;
 }
 
 export const TravelDateTextInputFields: React.FC<
@@ -30,6 +34,10 @@ export const TravelDateTextInputFields: React.FC<
   endDateLabel = "To",
   onFocus,
   calendarSize,
+  placeholderWhenBlurredStartDate,
+  placeholderWhenBlurredEndDate,
+  valueWhenBlurredStartDate,
+  valueWhenBlurredEndDate,
 }) => {
   const { mask, placeholder } = useMemo(() => {
     const dateFormatForLocaleCode = getDateFormatForLocaleCode(localeCode);
@@ -60,6 +68,8 @@ export const TravelDateTextInputFields: React.FC<
         label={startDateLabel}
         borderRadiusVariant={"onlyLeft"}
         placeholder={placeholder}
+        placeholderWhenBlurred={placeholderWhenBlurredStartDate}
+        valueWhenBlurred={valueWhenBlurredStartDate}
         calendarSize={calendarSize}
       />
       <TravelDateTextInput
@@ -79,6 +89,8 @@ export const TravelDateTextInputFields: React.FC<
         label={endDateLabel}
         borderRadiusVariant={"onlyRight"}
         placeholder={placeholder}
+        placeholderWhenBlurred={placeholderWhenBlurredEndDate}
+        valueWhenBlurred={valueWhenBlurredEndDate}
         calendarSize={calendarSize}
       />
     </Row>