+++ packages/tooltip/src/types.ts
@@ -1,3 +1,0 @@
-import { Placement } from "tippy.js";
-
-export type PopoverPlacement = Placement;